Add MenuItemContainerTrait (task #809)
