+++ src/MenuBuilder/Menu.php
@@ -11,15 +11,14 @@
  */
 namespace Menu\MenuBuilder;
 
-use Menu\MenuBuilder\MenuItem;
-use Menu\MenuBuilder\MenuItemRenderFactory;
-
 /**
  *  Menu class
  *
  */
 class Menu implements MenuInterface
 {
+    use MenuItemContainerTrait;
+
     /**
      * @const MENU_BUTTONS_TYPE
      */
@@ -31,11 +30,6 @@ class Menu implements MenuInterface
     const MENU_ACTIONS_TYPE = 'actions';
 
     /**
-     * @var $menuItems
-     */
-    protected $menuItems = [];
-
-    /**
      *  addMenuItem method
      *
      * @param \Menu\MenuBuilder\MenuItemInterface $item menu item definition
@@ -43,7 +37,7 @@ class Menu implements MenuInterface
      */
     public function addMenuItem($item)
     {
-        array_push($this->menuItems, $item);
+        $this->add($item);
     }
 
     /**
@@ -53,46 +47,6 @@ class Menu implements MenuInterface
      */
     public function getMenuItems()
     {
-        usort($this->menuItems, function ($a, $b) {
-            return $a->getOrder() > $b->getOrder();
-        });
-
-        return $this->menuItems;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param MenuItemInterface $item the menu item to be added
-     * @return void
-     */
-    public function add(MenuItemInterface $item)
-    {
-        $this->addMenuItem($item);
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @return array List of menu items
-     */
-    public function getAll()
-    {
-        return $this->getMenuItems();
-    }
-
-    /**
-     * Removes the specified menu item from this container.
-     * If the provided item is not in this container, this method does nothing.
-     *
-     * @param MenuItemInterface $item The item to be removed from the menu.
-     * @return void
-     */
-    public function remove(MenuItemInterface $item)
-    {
-        $key = array_search($item, $this->menuItems);
-        if ($key !== false) {
-            unset($this->menuItems[$key]);
-        }
+        return $this->getAll();
     }
 }
